Minor changes to neighbor list. Merge all functions together in optimize.
